CHROMIUM: provide default implementation for signal handlers
A D-Bus client may not be interested of all the D-Bus signals emitted by
a D-Bus server. This CL modifies the proxy template in dbus-c++ such
that the base proxy class provides a default implementation for signal
handlers, which can be overridden by a subclass as needed.

BUG=chromium:776452
TEST=Build cromo and wimax-manager.
TEST=Remote trybot runs.

Change-Id: I64057d31dc1784bddc29bbb02977fe9a29e1a81e
Reviewed-on: https://chromium-review.googlesource.com/917528
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>