Fix null pointer dereference in libdbus-c++
When the G-Lib mainloop notifies libdbus-c++ that an event has
happened on a BusWatch object (which D-Bus uses to integrate its
connection file descriptors with your event loop - G-Lib in our case),
there is a window where the GSource object has been nulled out. This
has caused several crashes in stackd.

The theory is that this can happen if D-Bus decides to re-cycle the
connection while data is available on the old connection.
Unfortunately this is very difficult to reproduce, so this change
merely handles this case by removing the dangling GSource object.